+++ example/main.dart
@@ -17,26 +17,85 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String url1 =
-      'https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg';
+  String url1 = 'https://www.sefram.com/images/products/photos/hi_res/7202.jpg';
+
+  bool isImageCached = false;
+  String? log;
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
         home: Scaffold(
-            body: SizedBox(
-      height: 350,
-      width: 350,
-      child: FastCachedImage(
-        url: url1,
-        fit: BoxFit.cover,
-        fadeInDuration: const Duration(seconds: 1),
-        errorBuilder: (context, exception, stacktrace) {
-          return Text(exception.toString());
-        },
-        loadingBuilder: (context) {
-          return Container(color: Colors.grey);
-        },
+            body: Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          SizedBox(
+            height: 150,
+            width: 150,
+            child: FastCachedImage(
+              url: url1,
+              fit: BoxFit.cover,
+              fadeInDuration: const Duration(seconds: 1),
+              errorBuilder: (context, exception, stacktrace) {
+                return Text(stacktrace.toString());
+              },
+              loadingBuilder: (context, progress) {
+                return Container(
+                  color: Colors.yellow,
+                  child: Stack(
+                    alignment: Alignment.center,
+                    children: [
+                      if (progress.isDownloading && progress.totalBytes != null)
+                        Text(
+                            '${progress.downloadedBytes ~/ 1024} / ${progress.totalBytes! ~/ 1024} kb',
+                            style: const TextStyle(color: Colors.red)),
+                      SizedBox(
+                          width: 120,
+                          height: 120,
+                          child: CircularProgressIndicator(
+                              color: Colors.red,
+                              value: progress.progressPercentage.value)),
+                    ],
+                  ),
+                );
+              },
+            ),
+          ),
+          const SizedBox(height: 12),
+          Text('Is image cached? = $isImageCached',
+              style: const TextStyle(color: Colors.red)),
+          const SizedBox(height: 12),
+          Text(log ?? ''),
+          const SizedBox(height: 120),
+          MaterialButton(
+            onPressed: () async {
+              setState(() => isImageCached =
+                  FastCachedImageConfig.isCached(imageUrl: url1));
+            },
+            child: const Text('check image is cached or not'),
+          ),
+          const SizedBox(height: 12),
+          MaterialButton(
+            onPressed: () async {
+              await FastCachedImageConfig.deleteCachedImage(imageUrl: url1);
+              setState(() => log = 'deleted image $url1');
+              await Future.delayed(
+                  const Duration(seconds: 2), () => setState(() => log = null));
+            },
+            child: const Text('delete cached image'),
+          ),
+          const SizedBox(height: 12),
+          MaterialButton(
+            onPressed: () async {
+              await FastCachedImageConfig.clearAllCachedImages();
+              setState(() => log = 'All cached images deleted');
+              await Future.delayed(
+                  const Duration(seconds: 2), () => setState(() => log = null));
+            },
+            child: const Text('delete all cached images'),
+          ),
+        ],
       ),
     )));
   }